Split mode visualization and profile basis helpers
