+++ beamz/devices/sources/mode.py
@@ -1,57 +1,36 @@
-import warnings
 from dataclasses import replace
 
-import jax.numpy as jnp
-import numpy as np
-
-from beamz.const import EPS_0, LIGHT_SPEED, MU_0
 from beamz.devices.sources import apply as apply_helpers
-from beamz.devices.sources.inject import (
-    _get_3d_huygens_terms,
-    _inject_3d_e_fields,
-    _inject_3d_fields,
-    _inject_3d_h_fields,
-    _inject_e_component,
-    _inject_h_component,
-    _match_shape,
-)
+from beamz.devices.sources import mode_visual
 from beamz.devices.sources.profiles import (
     _axis_index_from_component_indices,
-    _backward_3d_mode_from_forward,
-    _build_3d_profiles,
     _component_axis_coord,
     _dominant_3d_pair,
     _impedance_match_3d_tangential_pairs,
     _impedance_match_e_profile,
-    _make_3d_mode_basis_profiles,
-    _modal_overlap_3d_profiles,
-    _modal_power_2d,
-    _modal_power_3d_from_profiles,
-    _normalize_2d_pair_by_power,
-    _normalize_3d_profiles_by_flux,
     _numeric_impedance_axis,
     _numeric_phase_delay,
     _parse_direction,
-    _project_3d_profiles_to_real,
     _remap_3d_solver_components,
     _select_3d_impedance_index,
     _select_3d_phase_ref,
     _select_core_confined_mode_index,
     _solve_numeric_k_axis,
 )
+from beamz.devices.sources.profiles_basis import (
+    _backward_3d_mode_from_forward,
+    _make_3d_mode_basis_profiles,
+    _modal_overlap_3d_profiles,
+    _modal_power_2d,
+    _modal_power_3d_from_profiles,
+    _normalize_2d_pair_by_power,
+    _normalize_3d_profiles_by_flux,
+    _project_3d_profiles_to_real,
+)
 from beamz.devices.sources import setup as setup_helpers
 from beamz.devices.sources.spec import ModeSourceSpec, build_mode_source_spec
 from beamz.devices.sources.state import ModeSourceState
 from beamz.devices.sources.solve import solve_modes
-from beamz.devices.sources.windows import (
-    _compute_transverse_bounds,
-    _crop_and_window_2d,
-    _jax_tukey_window,
-    _make_tukey_window_2d,
-    _scipy_tukey,
-    _stagger_both,
-    _stagger_half,
-)
 
 
 # ---------------------------------------------------------------------------
@@ -100,6 +79,12 @@ _MODE_STATE_MAP = {
     "_axis": "axis",
     "_transverse_start": "transverse_start",
     "_transverse_end": "transverse_end",
+    "_x_start": "x_start",
+    "_x_end": "x_end",
+    "_y_start": "y_start",
+    "_y_end": "y_end",
+    "_z_start": "z_start",
+    "_z_end": "z_end",
 }
 
 
@@ -231,92 +216,17 @@ class ModeSource:
 
     def profile_data(self, field=None):
         """Return the current source profile as plotting-ready data."""
-        from beamz.visual.data import Slice2D, Trace1D
-
-        if self._Ez_profile is None and self._jz_profile is None:
-            if self.grid is not None and hasattr(self.grid, "permittivity"):
-                resolution = getattr(self.grid, "resolution", 0.05e-6)
-                self.initialize(self.grid.permittivity, resolution)
-            else:
-                raise ValueError(
-                    "Mode source is not initialized and no grid permittivity is available."
-                )
-
-        choices = {
-            "ez": ("Ez", self._Ez_profile),
-            "hz": ("Hz", self._jz_profile),
-            "jz": ("Hz", self._jz_profile),
-        }
-        key = None if field is None else str(field).strip().lower()
-        if key is None:
-            label, profile = ("Ez", self._Ez_profile)
-            if profile is None:
-                label, profile = ("Hz", self._jz_profile)
-        elif key in choices:
-            label, profile = choices[key]
-        else:
-            raise ValueError("field must be one of None, 'Ez', 'Hz', or 'Jz'.")
-
-        if profile is None:
-            raise ValueError(f"No profile data available for field '{field}'.")
-
-        profile = np.squeeze(np.asarray(profile))
-        title = f"{label} mode profile"
-        if self._neff is not None:
-            title = f"{title} (neff={self._neff:.4f})"
-
-        if profile.ndim == 2:
-            if self.direction in {"+x", "-x"}:
-                plane, x_label, y_label = "yz", "Y index", "Z index"
-            else:
-                plane, x_label, y_label = "xz", "X index", "Z index"
-            height, width = profile.shape
-            return Slice2D(
-                values=profile,
-                extent=(0.0, float(max(width - 1, 1)), 0.0, float(max(height - 1, 1))),
-                value_label="Amplitude",
-                plane=plane,
-                title=title,
-                x_label=x_label,
-                y_label=y_label,
-                style={"cmap": "magma", "origin": "lower", "aspect": "auto"},
-            )
-
-        return Trace1D(
-            values=profile.reshape(-1),
-            coords=np.arange(profile.size, dtype=float),
-            coord_label="index",
-            value_label="Amplitude",
-            title=title,
-            style={"color": "black", "linewidth": 2},
-        )
+        return mode_visual.profile_data(self, field=field)
 
     def show(self, field=None):
         """Visualize the 2D mode profile (for 3D simulations) or 1D profile (for 2D)."""
-        import matplotlib.pyplot as plt
-        from beamz.visual.data import Slice2D
-
-        try:
-            plot_data = self.profile_data(field=field)
-        except ValueError as exc:
-            warnings.warn(str(exc), stacklevel=2)
-            return None
-        if isinstance(plot_data, Slice2D):
-            plot_data.plot(cmap="magma", abs_value=True, aspect="auto")
-        else:
-            ax = plot_data.plot(color="k", abs_value=True)
-            ax.grid(True)
-        plt.tight_layout()
-        plt.show()
-        return plot_data
+        return mode_visual.show(self, field=field)
 
     def add_to_plot(
         self, ax, facecolor="none", edgecolor="crimson", alpha=0.8, linestyle="-"
     ):
         """Add source visualization to 2D matplotlib plot."""
-        from beamz.visual.overlays import add_mode_source_to_plot
-
-        add_mode_source_to_plot(
+        mode_visual.add_to_plot(
             self,
             ax,
             facecolor=facecolor,
